:sparkles: add theme switching and first custom theme
Now the app is switching between the different themes in ThemeData.swift, when the UserDefault Theme changes. Actually it needs to reopen the window that the changes take effect, but maybe there is a smarter way in the future. I already created a custom theme that should be similar to solarized. The ThemeData struct isn't completed yet, I plan to add other props, f.e. some for the fonts, in the future.